Trying different approach with constructor functions
Using constructor functions instead of regular functions, developers can create inheritance models with their decorators.

Also I didn't understand the value in generating a static chain of responsibility by returning the descriptor in the decorator function. As the descriptor is an object reference, a decorator could simple decide if he'd like to modify the descriptor or not before it will be defined on the object.